# swagger.yml
swagger: "2.0"
info:
  title: "Transfer money"
  description: "Test task implementation for transfer money between accounts."
  version: "1.0.0"
host: "localhost:8080"
basePath: "/banking"
schemes:
  - "http"
paths:
  /accounts:
    get:
      summary: "Gets all accounts"
      description: ""
      produces:
        - "application/json"
      responses:
        200:
          description: "OK"
          schema:
            type: array
            items:
              $ref: "#/definitions/Account"
        500:
          description: "Server error"
  /accounts/{id}:
    get:
      summary: "Finds account by ID"
      description: "Returns single account"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of account"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Account"
        404:
          description: "account not found"

  /accounts/{id}/transfer:
    post:
      summary: "Performs money transfer from current account to target"
      description: ""
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of source account"
          required: true
          type: "integer"
          format: "int64"
        - name: "target_account_id"
          in: "query"
          description: "ID of target account"
          required: true
          type: "integer"
          format: "int64"
        - name: "amount"
          in: "query"
          description: "Amount to transfer"
          required: true
          type: "number"
          format: "number"
      responses:
        204:
          description: "OK"
        404:
          description: "Source or target account not found"
        422:
          description: "Invalid parameters"
        500:
          description: "Server error"
  /accounts/{id}/transactions:
    get:
      summary: "Finds transactions of particular account ID"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of account"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "OK"
          schema:
            type: array
            items:
              $ref: "#/definitions/Transaction"
        404:
          description: "account not found"
        500:
          description: "Server error"
definitions:
  Account:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      balance:
        type: "number"
        format: "number"
  Transaction:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      transfer_id:
        type: "string"
      amount:
        type: "number"
      date:
        type: "string"